refactor: move local search route to seperate file
